fix: initialize release notes widget with hook AYC-62
