+++ ayunis-core-frontend/src/widgets/app-sidebar/ui/AppSidebar.tsx
@@ -41,6 +41,7 @@ import { useTheme } from '@/features/theme';
 import { useSidebar } from '@/shared/ui/shadcn/sidebar';
 import { MeResponseDtoSystemRole } from '@/shared/api/generated/ayunisCoreAPI.schemas';
 import config from '@/shared/config';
+import { useReleaseNotes } from '@/features/useReleaseNotes';
 
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   const { theme } = useTheme();
@@ -49,6 +50,7 @@ export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   const { t } = useTranslation('common');
   const navigate = useNavigate();
   const { closeMobileWithCleanup } = useSidebar();
+  useReleaseNotes();
 
   useKeyboardShortcut(['j', 'Meta'], () => {
     void navigate({ to: '/chat' });
